Adjust the step in the lyric importer.
