+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Import/Lyrics/GenerateRubyRomaji/GenerateRubyRomajiNavigation.cs
@@ -11,7 +11,6 @@ namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Import.Lyrics.GenerateRubyRomaj
 public partial class GenerateRubyRomajiNavigation : TopNavigation<GenerateRubyRomajiStepScreen>
 {
     private const string auto_generate_ruby = "AUTO_GENERATE_RUBY";
-    private const string auto_generate_romaji = "AUTO_GENERATE_ROMAJI";
 
     public GenerateRubyRomajiNavigation(GenerateRubyRomajiStepScreen screen)
         : base(screen)
@@ -43,9 +42,9 @@ public partial class GenerateRubyRomajiNavigation : TopNavigation<GenerateRubyRo
     protected override LocalisableString GetNavigationText(NavigationState value) =>
         value switch
         {
-            NavigationState.Initial => $"Lazy to typing ruby? Press [{auto_generate_ruby}] or [{auto_generate_romaji}] to auto-generate ruby and romaji. It's very easy.",
-            NavigationState.Working => $"Go to next step to generate time-tag. Messing around? Press [{auto_generate_ruby}] or [{auto_generate_romaji}] again.",
-            NavigationState.Done => $"Go to next step to generate time-tag. Messing around? Press [{auto_generate_ruby}] or [{auto_generate_romaji}] again.",
+            NavigationState.Initial => $"Lazy to typing ruby? Press [{auto_generate_ruby}].",
+            NavigationState.Working => $"Go to next step to generate time-tag.",
+            NavigationState.Done => $"Go to next step to generate time-tag.",
             NavigationState.Error => "Oops, seems cause some error in here.",
             _ => throw new ArgumentOutOfRangeException(nameof(value)),
         };
@@ -58,7 +57,6 @@ public partial class GenerateRubyRomajiNavigation : TopNavigation<GenerateRubyRo
         public GenerateRubyTextFlowContainer(GenerateRubyRomajiStepScreen screen)
         {
             AddLinkFactory(auto_generate_ruby, "auto generate ruby", screen.AskForAutoGenerateRuby);
-            AddLinkFactory(auto_generate_romaji, "auto generate romaji", screen.AskForAutoGenerateRomaji);
         }
     }
 }
